requirements.txt
awscli==1.22.34
awsiot==0.1.3
blinker==1.4
boto3==1.35.71
botocore==1.35.71
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
docutils==0.17.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
jmespath==1.0.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paho-mqtt==2.1.0
Pillow==9.0.1
pyasn1==0.4.8
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pyserial==3.5
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
PyYAML==5.4.1
requests==2.25.1
roman==3.3
rsa==4.8
s3transfer==0.10.4
SecretStorage==3.3.1
six==1.16.0
systemd-python==234
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.3
wadllib==1.3.6
zipp==1.0.0
